backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.18
psycopg2-binary==2.9.6
pydantic==1.10.9
alembic==1.11.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==1.3.1
python-dotenv==1.0.0
jinja2==3.1.2
python-dateutil==2.8.2
requests==2.31.0
twilio==8.12.0  # For SMS messaging 